requirements.txt
fastapi
uvicorn
python-dotenv
pytubefix
httpx
